refactor: use new getProvider and getSelectedNetwork helpers
